refactor(startup): reduce main.dart to an entry point
main.dart still held the whole launch path: a 272-line startup-coordinator
factory, a 599-line blocking bootstrap, nine startup phases, the shorebird
and build-provenance helpers, and the root app widget. Every one of them had
to be reached through the entry point, which is why so many of their tests
import `package:openvine/main.dart as app`.

Split along the seams that were already there:

- startup/timed_startup_task.dart — the timing wrapper both the coordinator
  and the bootstrap use.
- startup/startup_phases.dart — the nine phases, now public because two
  libraries legitimately call them.
- startup/startup_coordinator_factory.dart — the phase registration.
- startup/app_bootstrap.dart — the blocking sequence up to runApp.
- startup/window_size_constants.dart — the desktop window size.
- app/divine_app.dart — DivineApp, its state, and the quick-actions navigator.

The moves are mechanical, and verified rather than asserted: startOpenVineApp
(599 lines), createStartupCoordinator (272), initializeCoreServices,
initializeZendeskSupport, routeNotificationTap and handleUncaughtZoneError
each diff to zero against origin/main once the privacy renames are normalised
away.

The startup error widget stays in main.dart and is injected into
startOpenVineApp. It renders before any theme, font loader or Directionality
exists, so its raw TextStyles with an explicit TextDecoration.none are right
there — a VineTheme helper would start an async Google font fetch on a crash
screen. Injecting it also keeps app_bootstrap from importing the entry point.

Adds the guard the issue asks for: a ninth copy of the RouteType-to-tab map
anywhere under lib/ fails a test. It matches the switch shape all eight copies
had, and deliberately not the map literal in last_tab_position_provider, which
stores a remembered scroll index rather than a tab.

The root-PopScope guard moves with the MaterialApp: pointed at main.dart it
would have passed for the wrong reason, since main.dart can no longer contain
one. It now scans every library that builds or hosts the app root, and fails
if none of them mentions MaterialApp, so it cannot go vacuous again silently.

check_skip_ceiling.sh excludes video_cache_startup_test.dart by name — its
`skip:` is an argument to the function under test. The exclusion followed the
file to its new path.

main.dart: 3450 -> 82 lines, 154 -> 5 imports.
