+++ mobile/lib/startup/window_size_constants.dart
@@ -1,0 +1,11 @@
+// ABOUTME: Base desktop window dimensions for the vine aspect ratio
+// ABOUTME: Split out of main.dart so startup can size the window (#3337)
+
+/// Window size constants for desktop experience
+class WindowSizeConstants {
+  WindowSizeConstants._();
+
+  // Base dimensions for desktop vine experience (1x scale)
+  static const double baseWidth = 450;
+  static const double baseHeight = 700;
+}
