+++ mobile/test/startup/root_pop_scope_test.dart
@@ -1,4 +1,4 @@
-// ABOUTME: Guards main.dart against re-adding a PopScope above MaterialApp
+// ABOUTME: Guards the app root against re-adding a PopScope above MaterialApp
 // ABOUTME: Such a PopScope never registers a PopEntry, so its callback is dead
 
 import 'dart:io';
@@ -77,20 +77,46 @@ void main() {
       );
     });
 
-    test('is not reintroduced in main.dart', () {
-      final source = File('lib/main.dart').readAsStringSync();
+    // Scans the libraries that build or host the MaterialApp, which is where
+    // a root PopScope would land. Keep this list in step with whatever owns
+    // MaterialApp.router — a guard pointed at a file that can no longer
+    // contain the pattern passes without proving anything.
+    const rootLibraries = [
+      'lib/main.dart',
+      'lib/startup/divine_material_app.dart',
+      'lib/startup/app_composition_root.dart',
+      'lib/app/divine_app.dart',
+    ];
 
+    test('is not reintroduced above MaterialApp', () {
+      // Fails loudly if the app root moves again and this list goes stale.
+      final materialAppOwners = rootLibraries
+          .where(
+            (path) => File(path).readAsStringSync().contains('MaterialApp'),
+          )
+          .toList();
       expect(
-        source,
-        isNot(contains('PopScope')),
+        materialAppOwners,
+        isNotEmpty,
         reason:
-            'A PopScope in main.dart sits above MaterialApp.router and is '
-            'dead (see the widget test above). App-level back interception '
-            'belongs in resolveBackAction, reached through '
-            'BackButtonHandler on Android and the shell app bar elsewhere. '
-            'If you genuinely need to intercept above a Router, the '
-            'supported widget is NavigatorPopHandler.',
+            'None of $rootLibraries builds a MaterialApp any more, so this '
+            'guard is watching the wrong files. Point it at whatever does.',
       );
+
+      for (final path in rootLibraries) {
+        expect(
+          File(path).readAsStringSync(),
+          isNot(contains('PopScope')),
+          reason:
+              'A PopScope above MaterialApp.router is dead (see the widget '
+              'test above): it registers through ModalRoute.of(context), '
+              'which is null there. App-level back interception belongs in '
+              'resolveBackAction, reached through BackButtonHandler on '
+              'Android and the shell app bar elsewhere. To intercept above a '
+              'Router the supported widget is NavigatorPopHandler. Offending '
+              'file: $path',
+        );
+      }
     });
   });
 }
